#664: Move Sayg actions to dedicated controller
