+++ CourageScores/Controllers/GameController.cs
@@ -1,9 +1,6 @@
 ﻿using System.Diagnostics.CodeAnalysis;
-using CourageScores.Models.Cosmos.Game.Sayg;
 using CourageScores.Models.Dtos;
 using CourageScores.Models.Dtos.Game;
-using CourageScores.Models.Dtos.Game.Sayg;
-using CourageScores.Services;
 using CourageScores.Services.Command;
 using CourageScores.Services.Game;
 using Microsoft.AspNetCore.Mvc;
@@ -16,13 +13,11 @@ public class GameController : Controller
 {
     private readonly ICommandFactory _commandFactory;
     private readonly IGameService _gameService;
-    private readonly IGenericDataService<RecordedScoreAsYouGo, RecordedScoreAsYouGoDto> _saygStorageService;
 
-    public GameController(IGameService gameService, ICommandFactory commandFactory, IGenericDataService<RecordedScoreAsYouGo, RecordedScoreAsYouGoDto> saygStorageService)
+    public GameController(IGameService gameService, ICommandFactory commandFactory)
     {
         _gameService = gameService;
         _commandFactory = commandFactory;
-        _saygStorageService = saygStorageService;
     }
 
     [HttpGet("/api/Game/{id}")]
@@ -55,24 +50,5 @@ public class GameController : Controller
     public async Task<ActionResultDto<GameDto>> DeleteGame(Guid id, CancellationToken token)
     {
         return await _gameService.Delete(id, token);
-    }
-
-    [HttpPost("/api/Game/Sayg")]
-    public async Task<ActionResultDto<RecordedScoreAsYouGoDto>> StoreSaygData([FromBody] UpdateRecordedScoreAsYouGoDto data, CancellationToken token)
-    {
-        var command = _commandFactory.GetCommand<AddOrUpdateSaygCommand>().WithData(data);
-        return await _saygStorageService.Upsert(data.Id, command, token);
-    }
-
-    [HttpGet("/api/Game/Sayg/{id}")]
-    public async Task<RecordedScoreAsYouGoDto?> GetSaygData(Guid id, CancellationToken token)
-    {
-        return await _saygStorageService.Get(id, token);
-    }
-
-    [HttpDelete("/api/Game/Sayg/{id}")]
-    public async Task<ActionResultDto<RecordedScoreAsYouGoDto>> DeleteSaygData(Guid id, CancellationToken token)
-    {
-        return await _saygStorageService.Delete(id, token);
     }
 }
